Move entity message containers into entity.go.
